+++ extensions/CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - `generateLockfiles` parameter for `parseLockfile` to inhibit lockfile generation
+- `sandboxGeneration` parameter for `parseLockfile` to disable the lockfile
+    generation sandbox
 
 ### Fixed